requirements.txt
#All experimentations done on python 3.9.6

# dependencies for "PyPDF2"
PyPDF2

# dependencies for "py_pdf_parser"
py_pdf_parser
shapely
pyvoronoi
pandas
wheel
matplotlib

# dependencies for "LlamaParser"
llama-parse
python-dotenv
nest_asyncio

# dependencies for 'PyMUPDF"
pymupdf
ghostscript

# dependencies for 'Camelot"
camelot-py[base]
opencv-python
opencv-contrib-python
opencv-python-headless
opencv-contrib-python-headless

# dependencies for "Tabula"
tabula-py

# dependencies for document coverage metrics 
sentence-transformers
nltk
docx

# some other dependencies
pytesseract
opencv-python
opencv-python-headless
transformers
torch
spacy
SentencePiece
accelerate

# necessary import for llama-parse implementation
# import nest_asyncio
# nest_asyncio.apply()

